// src/models/dto/supplier-update-input.ts
export type SupplierUpdateInputDto = {
  legalName: string;
  tradeName?: string;
  document?: string;
  email?: string;
  phone?: string;
  whatsapp?: string;
  contactName?: string;
  addressZipCode?: string;
  addressStreet?: string;
  addressNumber?: string;
  addressDistrict?: string;
  addressCity?: string;
  addressState?: string;
  notes?: string;
  isActive?: boolean;
};